+++ src/Farmer/Arm/Bastion.fs
@@ -17,9 +17,9 @@ type BastionHost =
         member this.ResourceName = this.Name
         member this.JsonModel =
             let dependsOn = [
-                ArmExpression.resourceId(virtualNetworks, this.VirtualNetwork).Eval() |> ResourceName
+                ResourceId.create(virtualNetworks, this.VirtualNetwork)
                 for config in this.IpConfigs do
-                    ArmExpression.resourceId(publicIPAddresses, config.PublicIpName).Eval() |> ResourceName
+                    ResourceId.create (publicIPAddresses, config.PublicIpName)
             ]
             {| bastionHosts.Create(this.Name, this.Location, dependsOn, this.Tags) with
                 properties =
@@ -28,8 +28,8 @@ type BastionHost =
                            |> List.mapi(fun index ipConfig ->
                                {| name = sprintf "ipconfig%i" (index + 1)
                                   properties =
-                                   {| publicIPAddress = {| id = ArmExpression.resourceId(publicIPAddresses, ipConfig.PublicIpName).Eval() |}
-                                      subnet = {| id = ArmExpression.resourceId(subnets, this.VirtualNetwork, ResourceName "AzureBastionSubnet").Eval() |}
+                                   {| publicIPAddress = {| id = ResourceId.create(publicIPAddresses, ipConfig.PublicIpName).Eval() |}
+                                      subnet = {| id = ResourceId.create(subnets, this.VirtualNetwork, ResourceName "AzureBastionSubnet").Eval() |}
                                    |}
                                |})
                     |}