+++ src/Farmer/Builders/Builders.KeyVault.fs
@@ -46,7 +46,7 @@ type SecretConfig =
       Enabled : bool option
       ActivationDate : DateTime option
       ExpirationDate : DateTime option
-      Dependencies : ResourceName list }
+      Dependencies : ResourceId list }
     static member internal createUnsafe key =
         { Key = key
           Value = ParameterSecret(SecureParameter key)
@@ -92,7 +92,7 @@ type KeyVaultConfig =
       NetworkAcl : NetworkAcl
       Uri : Uri option
       Secrets : SecretConfig list
-      Dependencies : ResourceName list
+      Dependencies : ResourceId list
       Tags: Map<string,string>  }
       interface IBuilder with
         member this.DependencyName = this.Name
@@ -145,7 +145,7 @@ type KeyVaultConfig =
                   ActivationDate = secret.ActivationDate
                   ExpirationDate = secret.ExpirationDate
                   Location = location
-                  Dependencies = this.Name :: secret.Dependencies }
+                  Dependencies = ResourceId.create this.Name :: secret.Dependencies }
         ]
 
 type AccessPolicyBuilder() =
@@ -209,7 +209,6 @@ type KeyVaultBuilderState =
       Policies : AccessPolicyConfig list
       Uri : Uri option
       Secrets : SecretConfig list
-      Dependencies : ResourceName list
       Tags: Map<string,string> }
 
 type KeyVaultBuilder() =
@@ -217,13 +216,12 @@ type KeyVaultBuilder() =
         { Name = ResourceName.Empty
           TenantId = Subscription.TenantId
           Access = { VirtualMachineAccess = None; ResourceManagerAccess = Some Enabled; AzureDiskEncryptionAccess = None; SoftDelete = None }
-          Sku = Sku.Standard
+          Sku = Standard
           NetworkAcl = { IpRules = []; VnetRules = []; Bypass = None; DefaultAction = None }
           Policies = []
           CreateMode = None
           Uri = None
           Secrets = []
-          Dependencies = []
           Tags = Map.empty  }
 
     member __.Run(state:KeyVaultBuilderState) : KeyVaultConfig =
@@ -240,7 +238,10 @@ type KeyVaultBuilder() =
             | Some SimpleCreateMode.Recover, [] -> failwith "Setting the creation mode to Recover requires at least one access policy. Use the accessPolicy builder to create a policy, and add it to the vault configuration using add_access_policy."
           Secrets = state.Secrets
           Uri = state.Uri
-          Dependencies = state.Dependencies
+          Dependencies =
+            state.Policies
+            |> List.choose(fun r -> r.ObjectId.Owner)
+            |> List.distinct
           Tags = state.Tags  }
     /// Sets the name of the vault.
     [<CustomOperation "name">]
@@ -350,13 +351,17 @@ type SecretBuilder() =
     member __.ActivationDate(state:SecretConfig, activationDate) = { state with ActivationDate = Some activationDate }
     [<CustomOperation "expiration_date">]
     member __.ExpirationDate(state:SecretConfig, expirationDate) = { state with ExpirationDate = Some expirationDate }
+
+    member private _.AddDependency (state:SecretConfig, resourceName:ResourceName) = { state with Dependencies = ResourceId.create resourceName :: state.Dependencies }
+    member private _.AddDependencies (state:SecretConfig, resourceNames:ResourceName list) = { state with Dependencies = (resourceNames |> List.map ResourceId.create) @ state.Dependencies }
+    /// Sets a dependency for the web app.
     [<CustomOperation "depends_on">]
-    member __.DependsOn(state:SecretConfig, resourceName) = { state with Dependencies = resourceName :: state.Dependencies }
-    member __.DependsOn(state:SecretConfig, resources) = { state with Dependencies = List.concat [ resources; state.Dependencies ] }
-    member __.DependsOn(state:SecretConfig, builder:IBuilder) = { state with Dependencies = builder.DependencyName :: state.Dependencies }
-    member __.DependsOn(state:SecretConfig, builders:IBuilder list) = { state with Dependencies = List.concat [ builders |> List.map (fun x -> x.DependencyName); state.Dependencies ] }
-    member __.DependsOn(state:SecretConfig, resource:IArmResource) = { state with Dependencies = resource.ResourceName :: state.Dependencies }
-    member __.DependsOn(state:SecretConfig, resources:IArmResource list) = { state with Dependencies = List.concat [ resources |> List.map (fun x -> x.ResourceName); state.Dependencies ] }
+    member this.DependsOn(state:SecretConfig, resourceName) = this.AddDependency(state, resourceName)
+    member this.DependsOn(state:SecretConfig, resources) = this.AddDependencies(state, resources)
+    member this.DependsOn(state:SecretConfig, builder:IBuilder) = this.AddDependency(state, builder.DependencyName)
+    member this.DependsOn(state:SecretConfig, builders:IBuilder list) = this.AddDependencies(state, builders |> List.map (fun x -> x.DependencyName))
+    member this.DependsOn(state:SecretConfig, resource:IArmResource) = this.AddDependency(state, resource.ResourceName)
+    member this.DependsOn(state:SecretConfig, resources:IArmResource list) = this.AddDependencies(state, resources |> List.map (fun x -> x.ResourceName))
 
 let secret = SecretBuilder()
 let keyVault = KeyVaultBuilder()