+++ src/Tests/AllTests.fs
@@ -10,11 +10,14 @@ let allTests =
         testList "All Tests" [
             testList "Builders" [
                 LogAnalytics.tests
+                AppInsights.tests
                 Bastion.tests
                 Cdn.tests
                 CognitiveServices.tests
                 ContainerGroup.tests
                 ContainerService.tests
+                Dns.tests
+                EventHub.tests
                 IotHub.tests
                 Storage.tests
                 ContainerRegistry.tests