+++ src/Tests/Template.fs
@@ -133,14 +133,14 @@ let tests = testList "Template" [
         let a = storageAccount { name "aaa" }
         let b = webApp { name "b"; depends_on a.Name.ResourceName }
 
-        Expect.equal b.Dependencies [ ResourceName "aaa" ] "Dependency should have been set"
+        Expect.equal b.Dependencies [ ResourceId.create (ResourceName "aaa") ] "Dependency should have been set"
     }
 
     test "Can add dependency through IBuilder" {
         let a = storageAccount { name "aaa" }
         let b = webApp { name "b"; depends_on a }
 
-        Expect.equal b.Dependencies [ ResourceName "aaa" ] "Dependency should have been set"
+        Expect.equal b.Dependencies [ ResourceId.create (ResourceName "aaa") ] "Dependency should have been set"
     }
 
     test "Can add dependencies through Resource Name" {
@@ -148,7 +148,7 @@ let tests = testList "Template" [
         let b = storageAccount { name "bbb" }
         let b = webApp { name "b"; depends_on [ a.Name.ResourceName; b.Name.ResourceName ] }
 
-        Expect.equal b.Dependencies [ ResourceName "aaa"; ResourceName "bbb" ] "Dependencies should have been set"
+        Expect.equal b.Dependencies [ ResourceId.create (ResourceName "aaa"); ResourceId.create (ResourceName "bbb") ] "Dependencies should have been set"
     }
 
     test "Can add dependencies through IBuilder" {
@@ -156,25 +156,31 @@ let tests = testList "Template" [
         let b = storageAccount { name "bbb" }
         let b = webApp { name "b"; depends_on [ a :> IBuilder; b :> IBuilder ] }
 
-        Expect.equal b.Dependencies [ ResourceName "aaa"; ResourceName "bbb" ] "Dependencies should have been set"
+        Expect.equal b.Dependencies [ ResourceId.create (ResourceName "aaa"); ResourceId.create (ResourceName "bbb") ] "Dependencies should have been set"
     }
 
-    test "Generates ARM Id with name only" {
-        let rid = ArmExpression.resourceId (Farmer.Arm.Network.connections, ResourceName "test")
+    test "Generates untyped Resource Id" {
+        let rid = ResourceId.create (ResourceName "test")
+        let id = rid.Eval()
+        Expect.equal id "[string('test')]" "resourceId template function should match"
+    }
+
+    test "Generates typed Resource Id" {
+        let rid = ResourceId.create (Arm.Network.connections, ResourceName "test")
         let id = rid.Eval()
         Expect.equal id "[resourceId('Microsoft.Network/connections', 'test')]" "resourceId template function should match"
     }
 
-    test "Generates ARM Id with name and group" {
-        let rid = ArmExpression.resourceId (Farmer.Arm.Network.connections, ResourceName "test", "myGroup")
+    test "Generates typed Resource Id with group" {
+        let rid = ResourceId.create (Arm.Network.connections, ResourceName "test", "myGroup")
         let id = rid.Eval()
         Expect.equal id "[resourceId('myGroup', 'Microsoft.Network/connections', 'test')]" "resourceId template function should match"
     }
 
-    test "Generates ARM Id with name, group, and subscription" {
-        let rid = ArmExpression.resourceId (Farmer.Arm.Network.connections, ResourceName "test", "myGroup", "5ed984d9-9e7e-4550-b73b-7af020a7620d")
+    test "Generates typed Resource Id with segments" {
+        let rid = ResourceId.create (Arm.Network.connections, ResourceName "test", ResourceName "segment1", ResourceName "segment2")
         let id = rid.Eval()
-        Expect.equal id "[resourceId('5ed984d9-9e7e-4550-b73b-7af020a7620d', 'myGroup', 'Microsoft.Network/connections', 'test')]" "resourceId template function should match"
+        Expect.equal id "[resourceId('Microsoft.Network/connections', 'test', 'segment1', 'segment2')]" "resourceId template function should match"
     }
 
     test "Fails if ARM expression is already quoted" {